Update the promotions form for a new FPDF and PHP 8
This updates the promotions form to work under PHP 8 and with the newest version of FPDF which is version 1.86

PHP 8 removed the get_magic_quotes_gpc() function and so this commit removes use of it

FPDF 1.8 converted most of the attributes of the pdf object to protected attributes. This commit changes the few cases where we were reading from those protected attributes and now uses other approaches to getting those values.